igt-gpu-tools: re-add version 1.22
When upstream upgraded to version 1.23 they added a `package.use.force`
entry that forces all use flags enabled, because disabling any of them
breaks the binary and/or build.  Unfortunately, they didn't keep the old
version around, so it's impossible to use without pulling in a bunch of
extra deps.  This commit brings back the old version until the newer
versions work with USE flags.