profiles: enable ModemManager for NetworkManager
NetworkManager requires ModemManager to bring up Bluetooth network
interfaces, so this commit re-enables it for `net-misc/networkmanager`.
To keep dependencies to a mimimum, it also disables the current default
modem driver for ModemManager, `qmi`.  Notably, it also does NOT enable
ModemManager for `kde-plasma/plasma-nm`, because that feature is useless
without a real modem.