profiles: Rethinking Qt component USE flags
Some of the Qt component USE flags really increase compile time and I
don't necessiarly need the features they enable.  In particular, most of
the PyQt flags are only needed for PyQt development, as are the Qt
Designer flags.  This commit disables them and moves the geolocation
flags into the laptop profile.