profiles: Add arm64 server profiles, rework plasma
This is a major overhaul of the profiles to include a new _server_ tree
to support 64-bit Rasberry Pis.  To keep things DRY, it moves a lot of
the base configuration of the Plasma profile into a new shared _base_
profile.  Additionally, a lot of new configuration is added to the
_base_, _server_, and _server/arm64_ source profiles to support the new
leaf profiles, `_rpi3_ and _rpi4_.