profiles/arm64: Revert ACCEPT_KEYWORDS to upstream
In the early days of arm64 it was easier to set ACCEPT_KEYWORDS to allow
all testing packages rather than setting them individually.  This has
proven more costly in update time than it's worth, so this commit
removes that setting from make.defaults and sets a few
package.accept_keywords for any existing packages that are not yet
stable.