profiles: Set USE flag `verify-sig` globally
The USE flag `verify-sig` tells portage to verify the upstream
cyptographic signature of packages that support it.  This is a nice
extra layer of supply-chain security.  Prior to this it was set
per-package on one, but as it is gaining support in more packages, and
defined in `use.desc` not `use.local.desc`, it is more appropriately set
globally.  This commit sets it globally in the `custom/base` profile.