profiles/base: Add modules-compress to make.defaults
This commit adds the `modules-compress` USE flag to the base profile
`make.defaults`.  It's added as a global default, because most
out-of-tree modules are loaded relatively late in the boot process and
their load speed is not critical.  Modules that load ealier should be
excluded via `package.use` negation, as this commit does for
`sys-fs/zfs-kmod`.