Remove `modules-compress` USE flag
The `modules-compress` USE flag is not especially useful since A) it can
actually slow down a compressed initramfs due to double compression and
forking overhead, and B) I use ZFS with compression by default.  A
recent update in how modules are signed also conflicts with it, so this
commit disables it for all packages.